Adding testing for all classes
Still have to implement the input and output.

I am considering changing the structure of the program to have the
integrator as part of the equation. Equations are likely to change,
the integrator to solve the equation isn't. It also allows me to have a
solve class for each equation.

Still also need to reimplement the search algorithms, there is something
weird going on with the running time of these in the old implementation
of the code which needs to be fixed. For some reason they are not
finishing.